fix: Mark all as visited with IndexedDB
Closes #26
